requirements.txt
feedparser
python-dotenv
supabase
beautifulsoup4
newspaper3k
streamlit # If deploying this dashboard